driver_nmea0183: Fix 32/64 bug in cycle enders.
Long standing bug, that became obvious when more NMEA decodes
were added.  A 32 bit mask used to index into a 40 member array.o
Also fix over agressive mode setting on $PSRFEPE.

+++ gpsd.h
@@ -582,8 +582,8 @@ struct gps_device_t {
 	 */
 	timestamp_t this_frac_time, last_frac_time;
 	bool latch_frac_time;
-	unsigned long lasttag;
-	unsigned long cycle_enders;
+	int lasttag;              /* index into nmea_phrase[] */
+	uint64_t cycle_enders;    /* bit map into nmea_phrase{} */
 	bool cycle_continue;
     } nmea;
 #endif /* NMEA0183_ENABLE */
